+++ cli/src/commands/index.ts
@@ -1,5 +1,6 @@
 export { buildProject } from './build.js';
 export { login } from './login.js';
+export { loginCloud } from './login/login-cloud.js';
 export { INewProjectOptions, newProject } from './new.js';
 export { IServeOptions, serveProject } from './serve.js';
 export { setConfiguration } from './set.js';